Update quantity data type in 'CartItem' class.
Because the quantity may have one decimal place. Also update the instance constructor argument datatype in CartServiceImpl.getFloorQuantitesInCart

+++ src/main/java/vidar/websystem/domain/CartItem.java
@@ -2,6 +2,8 @@ package vidar.websystem.domain;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author yishi.xing
  * @created Mar 3, 2024 - 6:45:10 PM
@@ -13,12 +15,12 @@ public class CartItem {
 	 * @param floor
 	 * @param quantity
 	 */
-	public CartItem(HardwoodFloor floor, Long quantity) {
+	public CartItem(HardwoodFloor floor, BigDecimal quantity) {
 		this.floor = floor;
 		this.quantity = quantity;
 	}
 
 	private HardwoodFloor floor;
 	
-	private Long quantity;
+	private BigDecimal quantity;
 }
